Fixed some logic in s21_sqrt.c

+++ src/s21_sqrt.c
@@ -2,14 +2,15 @@
 
 #include "s21_math.h"
 
+//Метод половиного деления
 long double s21_sqrt(double x) {
-  double result = 0;
+  long double result = 0;
 
   // Проверка на отрицательную
   if (x < 0)
     result = -S21_NAN;
-  else
-    result = (double)s21_pow(x, 0.5);
-
-  return result;
+  else 
+    result = s21_pow(x, 0.5);
+    
+  return (double)result;
 }
